# Speed up TestClearSocialInvitesOnAdd