--- a/go/systests/team_invite_test.go
+++ b/go/systests/team_invite_test.go
@@ -391,41 +391,53 @@ func TestClearSocialInvitesOnAdd(t *testing.T) {
 
 	// Disable gregor in this test so Ann does not immediately add Bob
 	// through SBS handler when bob proves Rooter.
-	ann := makeUserStandalone(t, "ann", standaloneUserArgs{disableGregor: true})
+	ann := makeUserStandalone(t, "ann", standaloneUserArgs{
+		disableGregor:            true,
+		suppressTeamChatAnnounce: true,
+	})
 	tt.users = append(tt.users, ann)
 
+	tracer := ann.tc.G.CTimeTracer(context.Background(), "test-tracer")
+	defer tracer.Finish()
+
+	tracer.Stage("bob")
 	bob := tt.addUser("bob")
 
+	tracer.Stage("team")
 	team := ann.createTeam()
 
 	t.Logf("Ann created team %q", team)
 
 	bobBadRooter := "other" + bob.username
 
+	tracer.Stage("add 1")
 	ann.addTeamMember(team, bob.username+"@rooter", keybase1.TeamRole_WRITER)
+	tracer.Stage("add 2")
 	ann.addTeamMember(team, bobBadRooter+"@rooter", keybase1.TeamRole_WRITER)
 
+	tracer.Stage("prove rooter")
 	bob.proveRooter()
 
 	// Because bob@rooter is now proven by bob, this will add bob as a
 	// member instead of making an invitation.
+	tracer.Stage("add 3")
 	ann.addTeamMember(team, bob.username+"@rooter", keybase1.TeamRole_WRITER)
 
+	tracer.Stage("get team")
 	t0, err := teams.GetTeamByNameForTest(context.TODO(), ann.tc.G, team, false, true)
 	require.NoError(t, err)
 
+	tracer.Stage("assertions")
 	writers, err := t0.UsersWithRole(keybase1.TeamRole_WRITER)
 	require.NoError(t, err)
 	require.Equal(t, len(writers), 1)
 	require.True(t, writers[0].Uid.Equal(bob.uid))
 
-	// Adding should have cleared bob...@rooter
 	hasInv, err := t0.HasActiveInvite(keybase1.TeamInviteName(bob.username), "rooter")
 	require.NoError(t, err)
-	require.False(t, hasInv)
+	require.False(t, hasInv, "Adding should have cleared bob...@rooter")
 
-	// But should not have cleared otherbob...@rooter
 	hasInv, err = t0.HasActiveInvite(keybase1.TeamInviteName(bobBadRooter), "rooter")
 	require.NoError(t, err)
-	require.True(t, hasInv)
+	require.True(t, hasInv, "But should not have cleared otherbob...@rooter")
 }