This test was taking 20s. This speeds it up to 7s.

The slowdown was chat failing to send team announcement messages, which it had no hope of because gregor was explicitly disabled for this test. This PR disabled the chat sends at the command runner level.

```
:chat-trace=SLlJQRwMDpUW]
        test_logger.go:66: ^M2017-11-13 15:19:55.35037 gregor.go:268: [D] ++Chat: | PushHandler: GetClient: shutdown, using OfflineClient for chat1.RemoteClient (waited 4s for connectHappened)
        test_logger.go:66: ^M2017-11-13 15:19:55.35044 helper.go:782: [D] ++Chat: - newConversationHelper: newConversationHelper -> ERROR: error preparing message: operation failed: no connection to chat server (8.008125064s) [tags:chat-
trace=SLlJQRwMDpUW]
        test_logger.go:66: ^M2017-11-13 15:19:55.35047 chatmessaging.go:16: [W] failed to send team welcome message: error preparing message: operation failed: no connection to chat server
        test_logger.go:66: ^M2017-11-13 15:19:55.35049 teams.go:54: [D] - TeamCreate(ttrdwvnnli) -> <nil> [time=8.300145366s]
        test_logger.go:66: ^M2017-11-13 15:19:55.35060 teams_test.go:231: [D] - tracer-create-team [time=8.300376789s]
```